Add files via upload
the php files are for the website .....check the website by (thepythons.rf.gd)  username: love password:love ...and dont finish my money in there because it is real linked with the cbe account. inside the website there is a pay method with specific id this data is linked with the arduino through the processing ide using serial port..... the code for the ardjuino part is inside the com file it is arduino file ... the code for the processing is inside the processig_take_data_from_serial file it is for the processing ide which can read data from the webserver and pass it through the serial port to the arduino.